[ REFACTOR ] 클라이언트 graphQL 로직 리팩토링 (#21)
Why
- deprecated된 외부 모듈을 사용중이다.
- graphQL 로직과 뷰 로직 분리가 필요하다.
- 반복되는 로직을 모듈화할 필요가 있다.

What
- Apollo-link-state 모듈을 삭제하고, apollo-client의 내장 기능 활용
- graphQL 로직을 커스텀 hook으로 분리하여 재활용함
