+++ web/package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "build": "webpack",
     "dev": "webpack-dev-server",
-    "test":""
+    "test": ""
   },
   "dependencies": {
     "@apollo/react-hooks": "^3.1.3",
@@ -18,9 +18,9 @@
     "apollo-link": "^1.2.13",
     "apollo-link-error": "^1.1.12",
     "apollo-link-http": "^1.5.16",
-    "apollo-link-state": "^0.4.2",
     "graphql": "^14.5.8",
     "graphql-tag": "^2.10.1",
+    "lodash": "^4.17.15",
     "prop-types": "^15.7.2",
     "react": "^16.11.0",
     "react-apollo": "^3.1.3",
