+++ web/src/graphql/link/index.js
@@ -2,20 +2,19 @@ import { from } from 'apollo-link';
 import { HttpLink } from 'apollo-link-http';
 import { onError } from 'apollo-link-error';
 import authLink from './auth';
-import stateLink from './state';
+import { GRAPHQL_API } from '@/constants';
 
 const errorLink = onError(({ graphQLErrors, networkError }) => {
   console.log(graphQLErrors, networkError);
 });
 
 const httpLink = new HttpLink({
-  uri: 'http://localhost:4000/graphql',
+  uri: GRAPHQL_API,
 });
 
 const link = from([
   errorLink,
   authLink,
-  stateLink,
   httpLink,
 ]);
