+++ web/src/graphql/resolvers/auth.js
@@ -1,15 +1,3 @@
-import { withClientState } from 'apollo-link-state';
-import inMemoryCache from '../cache';
-
-const defaults = {
-  authentication: {
-    __typename: 'authentication',
-    isLoggedIn: !!localStorage.getItem('DROPY_TOKEN'),
-    displayName: localStorage.getItem('DROPY_USERNAME'),
-    token: localStorage.getItem('DROPY_TOKEN'),
-  },
-};
-
 const resolvers = {
   Mutation: {
     logIn: (_, { token, displayName }, { cache }) => {
@@ -43,10 +31,4 @@ const resolvers = {
   },
 };
 
-const stateLink = withClientState({
-  defaults,
-  resolvers,
-  cache: inMemoryCache,
-});
-
-export default stateLink;
+export default resolvers;
