+++ web/src/graphql/resolvers/index.js
@@ -1,0 +1,4 @@
+import { merge } from 'lodash';
+import authResolvers from './auth';
+
+export default merge(authResolvers);
